pyproject.toml
[project]
name = "person-expense-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]  # Error, pyflakes, isort
ignore = ["E501"]  # Ignore line too long (formatter handles it)
